Protege os endpoints GET, PUT e DELETE com autenticação

+++ tests/test_app.py
@@ -75,7 +75,7 @@ def test_read_users_with_user(client, user):
 
 
 def test_read_user(client, user):
-    response = client.get('/users/1')
+    response = client.get(f'/users/{user.id}')
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {
@@ -85,27 +85,40 @@ def test_read_user(client, user):
     }
 
 
-def test_update_user(client, user):
-    response = client.put(
-        '/users/1',
-        json={
-            'username': 'Cuca',
-            'email': 'cuca@fdp.com',
-            'password': 'soufdp',
-        },
+def test_get_token(client, user):
+    response = client.post(
+        '/token', data={'username': user.email, 'password': user.clean_password}
     )
 
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert 'access_token' in token
+    assert token['token_type'] == 'Bearer'
+
+
+def test_update_user(client, user, token):
+    response = client.put(
+        f'/users/{user.id}',
+        headers={'Authorization': f'Bearer {token}'},
+        json={
+            'username': 'bob',
+            'email': 'bob@example.com',
+            'password': 'mynewpassword',
+        },
+    )
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {
-        'username': 'Cuca',
-        'email': 'cuca@fdp.com',
-        'id': 1,
+        'username': 'bob',
+        'email': 'bob@example.com',
+        'id': user.id,
     }
 
 
-def test_update_user_not_found(client):
+def test_update_user_not_found(client, token):
     response = client.put(
         '/users/-1',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'Cuca',
             'email': 'cuca@fdp.com',
@@ -113,20 +126,22 @@ def test_update_user_not_found(client):
         },
     )
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
+    assert response.status_code == HTTPStatus.FORBIDDEN
 
 
-def test_delete_user(client, user):
-    response = client.delete('/users/1')
+def test_delete_user(client, user, token):
+    response = client.delete(
+        f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'}
+    )
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'message': 'User deleted'}
 
 
-def test_delete_user_not_found(client):
-    response = client.delete('/users/-1')
+def test_delete_user_not_found(client, token):
+    response = client.delete('/users/-1', headers={'Authorization': f'Bearer {token}'})
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
+    assert response.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_read_user_not_found(client):
